requirements.txt
pynput==1.7.7
PyYAML==6.0.2
six==1.16.0
tinydb==4.8.2
